* Enlarge file.score field.

+++ system/db/chanzhi.sql
@@ -196,7 +196,7 @@ CREATE TABLE IF NOT EXISTS `eps_file` (
   `addedBy` char(30) NOT NULL DEFAULT '',
   `addedDate` datetime NOT NULL,
   `public` enum('1','0') NOT NULL DEFAULT '1',
-  `score` smallint unsigned NOT NULL DEFAULT 0,
+  `score` mediumint(8) unsigned NOT NULL DEFAULT 0,
   `downloads` mediumint(8) unsigned NOT NULL DEFAULT '0',
   `extra` varchar(255) NOT NULL,
   `primary` enum('1','0') DEFAULT '0',
